% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{delta.true}
\alias{delta.true}
\title{Calculate true interaction effects in the simulation study}
\usage{
delta.true(X, group, shift_func = shift)
}
\arguments{
\item{X}{A covariate matrix of subjects.}

\item{group}{A vector of group memberships of subjects.}

\item{shift_func}{A function for the method used in \code{equation (6)} to obtain
different true interaction effects for subjects in different groups.}
}
\value{
A matrix of true propensity scores.
Each row represents each subject.
Each column represents the interaction effects of covaraites and each treatment.
}
\description{
This function uses \code{equation (6)} in \verb{Section S.1.2} of the supplementary material
to calculate true interaction effects of covariates and treatments on the outcome
for subjects in the simulated datasets.
}
