% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlearn.R
\name{diff_reward}
\alias{diff_reward}
\title{Calculate the difference between rewards (outcomes)}
\usage{
diff_reward(data_MS, idx_MS, max_size = 1, delta, dist_mat)
}
\arguments{
\item{data_MS}{a matrix that contains outcome, treatment, and feature
information of a set of subjects.}

\item{idx_MS}{a data frame of two columns \code{ID} and \code{index}. \code{ID} records
the IDs of subjects in \code{data_MS}. \code{index} records the column indices of
these subjects in \code{dist_mat}.}

\item{max_size}{an integer indicating the upper limit of the sizes of all
matched sets. The default setting is \code{max_size=1} which means finding the
nearest neighbors of subjects.}

\item{delta}{a scalar, as defined in \code{equation (6)} in \verb{Section 2.2} of
the manuscript, indicating the upper limit of distances between a subject
and the subjects in its matched set. \cr
In future versions, we will extend this argument to a vector which means
the upper limit can vary with subjects.}

\item{dist_mat}{a precalculated matrix of distances between subjects.
This matrix must include all subjects in \code{data_MS}.}
}
\value{
A vector of length \code{max_size*nrow(data_MS)}.
The \code{(i-1)*max_size+1} to the \code{i*max_size} elements are the distances,
in ascending order, between the ith subject and the subjects in
its matched set. If the size of this matched set is smaller than \code{max_size},
some elements will be \code{NA}.
}
\description{
This function calculates the difference between outcomes of
a set of subjects and the subjects in their matched sets. That is, the
value of \code{Rj - Ri} in \code{equation (7)} in \verb{Section 2.2} of the manuscript.
}
