% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlearn.R
\name{evf}
\alias{evf}
\title{Calculate empirical value functions}
\usage{
evf(data, pred, propensity)
}
\arguments{
\item{data}{a data frame containing the ID (\code{ID}), outcome (\code{reward}),
and observed treatment (\code{treatment}) information of subjects.}

\item{pred}{a matrix with \code{K(K-1)/2} columns, where \code{K} is the total number
of treatments. Each column is the recommendations between two treatments
for all subjects.}

\item{propensity}{a data frame with \code{K} columns, where \code{K} is the total number
of treatments. The \code{k}th column is the propensity scores of assigning
the \code{k}th treatment to subjects.}
}
\value{
A scalar as defined in \code{equation (8)} in \verb{Section 2.2} of
the manuscript.
}
\description{
This function calculates the empirical value functions as
defined in \code{equation (8)} in \verb{Section 2.2} of the manuscript.
}
