% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlearn.R
\name{mlearn.wsvm}
\alias{mlearn.wsvm}
\title{Fit weighted kernel support vector machines (SVMs)}
\usage{
mlearn.wsvm(
  train_data,
  test_data,
  idx,
  trts,
  max_size,
  delta,
  dist_mat,
  g_func,
  kernel = "rbfdot",
  kpar = "automatic",
  C
)
}
\arguments{
\item{train_data}{a data frame for subjects in training fold(s), containing
the ID (\code{ID}), outcome (\code{reward}), outcome residual (\code{reward_res}), observed
treatment (\code{treatment}), and health feature information.}

\item{test_data}{a data frame for subjects in test fold(s), containing
the ID (\code{ID}), outcome (\code{reward}), outcome residual (\code{reward_res}), observed
treatment (\code{treatment}), and health feature information.}

\item{idx}{a data frame of two columns \code{ID} and \code{index}. \code{ID} records
the IDs of subjects in \code{train_data}. \code{index} records the column indices of
these subjects in \code{dist_mat}.}

\item{trts}{a vector of treatment names.}

\item{max_size}{an integer indicating the upper limit of the sizes of all
matched sets. The default setting is \code{max_size=1} which means finding the
nearest neighbors of subjects.}

\item{delta}{a scalar, as defined in \code{equation (6)} in \verb{Section 2.2} of
the manuscript, indicating the upper limit of distances between a subject
and the subjects in its matched set. \cr
In future versions, we will extend this argument to a vector which means
the upper limit can vary with subjects.}

\item{dist_mat}{a precalculated matrix of distances between subjects.
This matrix must include all subjects in \code{train_data}.}

\item{g_func}{a function that transforms the differences between outcomes
of a set of subjects and the subjects in their matched sets to the weights
in SVMs. In \code{equation (7)} in \verb{Section 2.2} of the manuscript,
\verb{g(.) = |.|} and the weights are \verb{|Rj-Ri|}.}

\item{kernel}{the kernel function used in SVMs. Supported argument values
can be found in \code{\link[kernlab]{ksvm}} and \code{\link[kernlab]{dots}}.
Default: "rbfdot".}

\item{kpar}{the list of hyper-parameters (kernel parameters). Valid
parameters for supported kernels can be found in \code{\link[kernlab]{ksvm}}
and \code{\link[kernlab]{dots}}. Default: "automatic".}

\item{C}{a scalar that is the cost of constraints violation in SVMs. This is
the "C"-constant of the regularization term in the Lagrange formulation.}
}
\value{
A list with 2 sublists as follows:
\itemize{
Suppose there are \code{K} treatments in the input data,
\item \code{model}: a list with \code{K(K-1)/2} sublists. Each sublist is a weighted
SVM (trained on \code{train_data}) for the corresponding binary treatment
comparison.
\item \code{prediction}: a matrix with \code{K(K-1)/2} columns. Each column is the
recommendations between the corresponding treatment pair for subjects in
\code{test_data}.
}
}
\description{
This function fits a set of weighted kernel SVMs
for binary treatment comparisons.
}
\seealso{
\code{\link[kernlab]{ksvm}} and \code{\link[kernlab]{dots}} for
\code{kernel}. \code{\link[personalized]{weighted.ksvm}} for fitting weighted SVMs.
}
