% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{mu.true}
\alias{mu.true}
\title{Calculate true main effects in the simulation study}
\usage{
mu.true(X)
}
\arguments{
\item{X}{a covariate matrix of subjects.}
}
\value{
A vector of true main effects of covariates.
}
\description{
This function uses \code{equation (5)} in \verb{Section S.1.2} of the supplementary material
to calculate true main effects of covariates on the outcome
for subjects in the simulated datasets.
}
