% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{pi.true}
\alias{pi.true}
\title{Calculate true propensity scores in the simulation study}
\usage{
pi.true(X, K)
}
\arguments{
\item{X}{a covariate matrix of subjects.}

\item{K}{the number of treatments.}
}
\value{
A matrix of true propensity scores of assigning each subject (row) to each treatment (column).
}
\description{
This function uses \code{equation (2)} in \verb{Section S.1.1} of the supplementary material
to calculate true propensity scores for subjects in the simulated datasets.
}
