% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{shift}
\alias{shift}
\title{Lead/lag for vectors and lists}
\usage{
shift(x, n = 1L, type = "lag")
}
\arguments{
\item{x}{a vector or list.}

\item{n}{an integer denoting the offset by which to lead or lag the input.
This function only supports n>=0.}

\item{type}{Default is \code{"lead"} (look \code{"forwards"}).
The other possible value is \code{"lag"} (look \code{"backwards"}).}
}
\value{
An object with the same class as \code{x} that contains the lead/lag of input \code{x}.
}
\description{
\code{Lead} or \code{lag} vectors or lists.
}
