% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simulate_data}
\alias{simulate_data}
\title{Generate a dataset for estimating individualized treatment rules}
\usage{
simulate_data(
  N = 200,
  p = 20,
  K = 4,
  J = 4,
  propensity_func,
  main_func,
  interaction_func
)
}
\arguments{
\item{N}{the number of subjects.}

\item{p}{the number of covariates.}

\item{K}{the number of treatments.}

\item{J}{the number of subject groups.}

\item{propensity_func}{a user-defined function that calculates true propensity scores of
assigning each subject to each treatment. \cr
An example of the accepted format of this function can be found in
\code{\link{pi.true}}.}

\item{main_func}{a user-defined function that calculates main effects of
covariates on the outcome for each subject. \cr
An example of the accepted format of this function can be found in
\code{\link{mu.true}}.}

\item{interaction_func}{a user-defined function that calculates interaction effects of
of covariates and treatments on the outcome for each subject. \cr
An example of the accepted format of this function can be found in
\code{\link{delta.true}}.}
}
\value{
A matrix containing the following columns of all subjects: \cr
\itemize{
\item \code{ID}: IDs.
\item \code{cluster}: group memberships.
\item \code{treatment}: observed/assigned treatments.
\item \code{reward}: values of the (continuous) outcome.
\item \code{feature_1}, \ldots, \code{feature_p}: values of the \code{p} covariates.
\item \code{interaction_1}, \ldots, \code{interaction_K}: values of interaction effects
of covariates and each of the \code{K} treatments.
\item \code{pi_1}, \ldots, \code{pi_K}: propensity scores of assigning each of the \code{K}
treatments.
\item \code{eps}: values of random errors.
\item \code{treatment_opt}: optimal treatments.
\item \code{reward_opt}: outcome values of the optimal treatments.
\item \code{pi_true}: propensity scores of assigning the observed treatments.
\item \code{pi_opt}: propensity scores of assigning the optimal treatments.
}
}
\description{
This function generates a simulated dataset for estimating
individualized treatment rules.
The outcome variable is assumed to follow \code{equation (1)} in
\verb{Section S.1.1} of the supplementary material.
}
\examples{
#######################################
## The simulated dataset in Section S.1.2 of
## the supplementary material for N=200

set.seed(0)
simdata200 = simulate_data(
  N = 200, p = 20, K = 4, J = 4,
  propensity_func = pi.true, # equation (2) in Section S.1.1
  main_func = mu.true, # equation (5) in Section S.1.2
  interaction_func = delta.true # equation (6) in Section S.1.2
)

}
\seealso{
\code{\link{pi.true}} for \code{propensity_func},
\code{\link{mu.true}} for \code{main_func},
and \code{\link{delta.true}} for \code{interaction_func}.
}
\concept{functions used in Section S.1 of the supplementary material}
