% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlearn.R
\name{summarize_rec}
\alias{summarize_rec}
\title{Summarize recommendations for multicategory treatment setup}
\usage{
summarize_rec(data, pred)
}
\arguments{
\item{data}{a data frame containing the ID (\code{ID}), outcome (\code{reward}),
and observed treatment (\code{treatment}) information of subjects.}

\item{pred}{a matrix with \code{K(K-1)/2} columns, where \code{K} is the total number
of treatments. Each column is the recommendations between two treatments
for all subjects.}
}
\value{
A data frame with the following 4 columns:
\itemize{
\item \code{ID}: IDs of subjects, same as those in \code{data}.
\item \code{reward}: outcome values of subjects, same as those in \code{data}.
\item \code{treatment}: observed treatments of subjects, same as those in \code{data}.
\item \code{vote}: recommended treatments of subjects that are summarized from \code{pred}
using the majority voting approach.
}
}
\description{
This function summarizes recommendations for multicategory
treatment setup from the results of binary treatment comparisons, using
the majority voting approach.
}
