% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlearn.R
\name{weighted.ksvm.2.cv}
\alias{weighted.ksvm.2.cv}
\title{Fit a nested cross-validation of weighted kernel support vector
machines (SVMs)}
\usage{
weighted.ksvm.2.cv(
  data,
  idx,
  trts,
  max_size,
  delta,
  dist_mat,
  g_func,
  kernel = "rbfdot",
  kpar = "automatic",
  nfolds_outer = 3,
  nfolds_inner = 3,
  tuneGrid,
  propensity,
  foldid_outer = NULL
)
}
\arguments{
\item{data}{a data frame containing the ID (\code{ID}), outcome (\code{reward_res}),
observed treatment (\code{treatment}), and health feature information of subjects.}

\item{idx}{a data frame of two columns \code{ID} and \code{index}. \code{ID} records
the IDs of subjects in \code{train_data}. \code{index} records the column indices of
these subjects in \code{dist_mat}.}

\item{trts}{a vector of treatment names.}

\item{max_size}{an integer indicating the upper limit of the sizes of all
matched sets. The default setting is \code{max_size=1} which means finding the
nearest neighbors of subjects.}

\item{delta}{a scalar, as defined in \code{equation (6)} in \verb{Section 2.2} of
the manuscript, indicating the upper limit of distances between a subject
and the subjects in its matched set. \cr
In future versions, we will extend this argument to a vector which means
the upper limit can vary with subjects.}

\item{dist_mat}{a precalculated matrix of distances between subjects.
This matrix must include all subjects in \code{train_data}.}

\item{g_func}{a function that transforms the differences between outcomes
of a set of subjects and the subjects in their matched sets to the weights
in SVMs. In \code{equation (6)} in \verb{Section 2.2} of the manuscript,
\verb{g(\\cdot) = |\\cdot|} and the weights are \verb{|Rj-Ri|}.}

\item{kernel}{the kernel function used in SVMs. Supported argument values
can be found in \code{\link[kernlab]{ksvm}} and \code{\link[kernlab]{dots}}.
Default: "rbfdot".}

\item{kpar}{the list of hyper-parameters (kernel parameters). Valid
parameters for supported kernels can be found in \code{\link[kernlab]{ksvm}}
and \code{\link[kernlab]{dots}}. Default: "automatic".}

\item{nfolds_outer}{the number of folds in the outer layer of the nested
cross-validation. Default: 3.}

\item{nfolds_inner}{the number of folds in the inner layer (for tuning
parameters) of the nested cross-validation. Values greater than or equal
to 3 usually yield better results. Default: 3.}

\item{tuneGrid}{a data frame of tuning parameter(s). Each column for each
parameter. Usually, the first column is the cost of constraints violation
("C"-constant) in SVMs.}

\item{propensity}{a data frame with \code{K} columns, where \code{K} is the total number
of treatments. The \code{k}th column is the propensity scores of assigning
the \code{k}th treatment to subjects.}

\item{foldit_outer}{(optional) a user-specified vector recording the split
of folds in the outer layer of the nested cross-validation. This vector
should match the number of rows in \code{data} and the number of treatments in
\code{trts}.}
}
\value{
A list with 3 sublists as follows:
\itemize{
\item \code{fit}: a list with \code{nfolds_outer} sublists. The \code{j}th sublist contains
the inner cross-validation result of the weighted SVM that used the \code{j}th
fold of subjects as the test fold.
\item \code{foldid_outer}: a vector recording the split of folds in the
outer layer of the nested cross-validation.
\item \code{prediction}: a matrix with 5 columns recording ID(\code{ID}),
outcome (\code{reward}), observed treatment (\code{treatment}), recommended
treatment (\code{vote}), and the fold in the cross-validation(\code{fold})
information of subjects.
}
}
\description{
This function performs a nested cross-validation
(on tuning parameters) of weighted SVMs for multicategory treatment
comparisons and estimating individualized treatment rules.
}
